Sort error logger entries
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@12724 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ plugins/electric-minarea-deltamerge0/src/main/java/com/sun/electric/plugins/minarea/deltamerge0/SimpleChecker.java
@@ -156,7 +156,7 @@ public class SimpleChecker implements MinAreaChecker {
             totalArea += larea;
             if (larea < minArea) {
                 Point2D p = poly.getPoints()[1];
-                errorLogger.reportMinAreaViolation(larea, DBMath.lambdaToGrid(p.getX()), DBMath.lambdaToGrid(p.getY()));
+                errorLogger.reportMinAreaViolation(larea, (int)DBMath.lambdaToGrid(p.getX()), (int)DBMath.lambdaToGrid(p.getY()));
             }
         }
         for (PolyBase.PolyBaseTree son : obj.getSons()) {
